refactor: split per-overlay hint bar helpers into their own file
Move overlayHintBarOverlayLocalClusters, overlayHintBarBookmarks, and
the per-overlay leaf hint helpers out of overlay_hintbar.go into a new
overlay_hintbar_overlays.go, keeping the routers and renderHints in
place and bringing overlay_hintbar.go back under the 800-line cap.
