+++ internal/app/overlay_hintbar_overlays.go
@@ -1,0 +1,441 @@
+package app
+
+import (
+	"github.com/janosmiko/lfk/internal/ui"
+)
+
+// overlayHintBarOverlayLocalClusters returns the hint bar entries for
+// the local-cluster manager. The screen state machine forks the keymap
+// across the list view, the five wizard steps, and the delete-confirm
+// sub-screen — each surface gets its own hint set so the bar matches
+// what the key handler actually consumes.
+func (m Model) overlayHintBarOverlayLocalClusters() string {
+	switch m.localClusterState.screen {
+	case localClusterScreenList:
+		if len(m.localClusterState.clusters) == 0 {
+			return m.renderHints([]ui.HintEntry{
+				{Key: "n", Desc: "new"},
+				{Key: "R", Desc: "refresh"},
+				{Key: "q/esc", Desc: "close"},
+			})
+		}
+		return m.renderHints([]ui.HintEntry{
+			{Key: "j/k", Desc: "navigate"},
+			{Key: "n", Desc: "new"},
+			{Key: "s", Desc: "start"},
+			{Key: "S", Desc: "stop"},
+			{Key: "D", Desc: "delete"},
+			{Key: "enter", Desc: "switch"},
+			{Key: "R", Desc: "refresh"},
+			{Key: "q/esc", Desc: "close"},
+		})
+	case localClusterScreenWizardProvider:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "j/k", Desc: "pick"},
+			{Key: "enter", Desc: "next"},
+			{Key: "esc", Desc: "back"},
+		})
+	case localClusterScreenWizardName,
+		localClusterScreenWizardVersion,
+		localClusterScreenWizardNodes:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "input"},
+			{Key: "enter", Desc: "next"},
+			{Key: "esc", Desc: "back"},
+		})
+	case localClusterScreenWizardConfirm:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "enter", Desc: "create"},
+			{Key: "esc", Desc: "back"},
+		})
+	case localClusterScreenDeleteConfirm:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type DELETE", Desc: "confirm"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	return ""
+}
+
+// overlayHintBarBookmarks returns hints for the bookmark overlay sub-modes.
+func (m Model) overlayHintBarBookmarks() string {
+	switch m.bookmarkSearchMode {
+	case bookmarkModeFilter:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "filter"},
+			{Key: "enter", Desc: "apply"},
+			{Key: "esc", Desc: "clear"},
+		})
+	case bookmarkModeConfirmDelete:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "Enter/y", Desc: "delete"},
+			{Key: "Esc/n", Desc: "cancel"},
+		})
+	case bookmarkModeConfirmDeleteAll:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "Enter/y", Desc: "delete all"},
+			{Key: "Esc/n", Desc: "cancel"},
+		})
+	default:
+		// tab desc names the action Tab will perform next: loading is the
+		// default, so it reads "don't load ns" until the user opts out, then
+		// "load ns". Matches the title chip the user sees.
+		tabDesc := "load ns"
+		if m.bookmarkLoadNamespace {
+			tabDesc = "don't load ns"
+		}
+		return m.renderHints([]ui.HintEntry{
+			{Key: "a-z/0-9", Desc: "jump"},
+			{Key: "enter", Desc: "jump"},
+			{Key: "tab", Desc: tabDesc},
+			{Key: "/", Desc: "filter"},
+			{Key: "ctrl+x", Desc: "delete"},
+			{Key: "alt+x", Desc: "delete all"},
+			{Key: "esc", Desc: "close"},
+		})
+	}
+}
+
+func (m Model) overlayHintBarOverlayLogContainerSelect() string {
+	hints := []ui.HintEntry{
+		{Key: "space", Desc: "select"},
+		{Key: "enter", Desc: "apply"},
+		{Key: "/", Desc: "filter"},
+	}
+	if m.logView.parentKind != "" {
+		hints = append(hints, ui.HintEntry{Key: "P", Desc: "switch pod"})
+	}
+	hints = append(hints, ui.HintEntry{Key: "esc", Desc: "close"})
+	return m.renderHints(hints)
+}
+
+func (m Model) overlayHintBarOverlayTemplates() string {
+	if m.templateSearchMode {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "filter"},
+			{Key: "enter", Desc: "apply"},
+			{Key: "esc", Desc: "clear"},
+		})
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "enter", Desc: "select"},
+		{Key: "/", Desc: "filter"},
+		{Key: "d", Desc: "delete saved"},
+		{Key: "esc", Desc: "close"},
+	})
+}
+
+func (m Model) overlayHintBarOverlayEventTimeline() string {
+	if m.eventTimelineSearchActive {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "search"},
+			{Key: "enter", Desc: "find"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	if m.eventTimelineVisualMode != 0 {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "j/k", Desc: "extend"},
+			{Key: "y", Desc: "copy"},
+			{Key: "v/V", Desc: "switch mode"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "j/k", Desc: "move"},
+		{Key: "g/G", Desc: "top/bottom"},
+		{Key: "v/V", Desc: "select"},
+		{Key: "y", Desc: "copy"},
+		{Key: ui.ActiveKeybindings.Search, Desc: "search"},
+		{Key: ui.ActiveKeybindings.Fullscreen, Desc: "fullscreen"},
+		{Key: ui.ActiveKeybindings.ToggleWrap, Desc: "wrap"},
+		{Key: "esc", Desc: "close"},
+	})
+}
+
+func (m Model) overlayHintBarOverlaySecretEditor() string {
+	if m.secretEditing {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "ctrl+s", Desc: "save"},
+			{Key: "enter", Desc: "newline"},
+			{Key: "tab", Desc: "switch"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	if m.editorSearch.active {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "filter"},
+			{Key: "enter", Desc: "apply"},
+			{Key: "esc", Desc: "clear"},
+		})
+	}
+	if m.editorSearch.formatActive {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "h/l", Desc: "format"},
+			{Key: "enter", Desc: "copy"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "jk", Desc: "nav"},
+		{Key: "v", Desc: "toggle"},
+		{Key: "V", Desc: "all"},
+		{Key: "e", Desc: "edit"},
+		{Key: "a", Desc: "add"},
+		{Key: "y", Desc: kvCopyHintDesc(m)},
+		{Key: "␣", Desc: "select"},
+		{Key: "Y", Desc: "copy as…"},
+		{Key: "/", Desc: "filter"},
+		{Key: "D", Desc: "del"},
+		{Key: "enter", Desc: "save"},
+		{Key: "esc", Desc: "close"},
+	})
+}
+
+// kvCopyHintDesc surfaces what `y` will do given the current
+// selection state — reflects the smart-y wiring (single-value copy
+// when no rows are marked, format-picker when selections are present).
+// Keeps the hint bar honest so the user doesn't press `y` and get a
+// surprise picker overlay.
+func kvCopyHintDesc(m Model) string {
+	if len(m.editorSearch.selected) > 0 {
+		return "copy as…"
+	}
+	return "copy"
+}
+
+func (m Model) overlayHintBarOverlayConfigMapEditor() string {
+	if m.configMapEditing {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "ctrl+s", Desc: "save"},
+			{Key: "enter", Desc: "newline"},
+			{Key: "tab", Desc: "switch"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	if m.editorSearch.active {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "filter"},
+			{Key: "enter", Desc: "apply"},
+			{Key: "esc", Desc: "clear"},
+		})
+	}
+	if m.editorSearch.formatActive {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "h/l", Desc: "format"},
+			{Key: "enter", Desc: "copy"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "jk", Desc: "nav"},
+		{Key: "e", Desc: "edit"},
+		{Key: "a", Desc: "add"},
+		{Key: "y", Desc: kvCopyHintDesc(m)},
+		{Key: "␣", Desc: "select"},
+		{Key: "Y", Desc: "copy as…"},
+		{Key: "/", Desc: "filter"},
+		{Key: "D", Desc: "del"},
+		{Key: "enter", Desc: "save"},
+		{Key: "esc", Desc: "close"},
+	})
+}
+
+func (m Model) overlayHintBarOverlayLabelEditor() string {
+	if m.labelEditing {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "ctrl+s", Desc: "save"},
+			{Key: "tab", Desc: "switch"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	if m.editorSearch.active {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "filter"},
+			{Key: "enter", Desc: "apply"},
+			{Key: "esc", Desc: "clear"},
+		})
+	}
+	if m.editorSearch.formatActive {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "h/l", Desc: "format"},
+			{Key: "enter", Desc: "copy"},
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "Tab", Desc: "switch"},
+		{Key: "jk", Desc: "nav"},
+		{Key: "e", Desc: "edit"},
+		{Key: "a", Desc: "add"},
+		{Key: "y", Desc: kvCopyHintDesc(m)},
+		{Key: "␣", Desc: "select"},
+		{Key: "Y", Desc: "copy as…"},
+		{Key: "/", Desc: "filter"},
+		{Key: "D", Desc: "del"},
+		{Key: "enter", Desc: "save"},
+		{Key: "esc", Desc: "close"},
+	})
+}
+
+func (m Model) overlayHintBarOverlayFinalizerSearch() string {
+	if m.finalizerSearchFilterActive {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "filter"},
+			{Key: "enter", Desc: "apply"},
+			{Key: "esc", Desc: "clear"},
+		})
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "space", Desc: "select"},
+		{Key: "ctrl+a", Desc: "all"},
+		{Key: "enter", Desc: "remove"},
+		{Key: "/", Desc: "filter"},
+		{Key: "esc", Desc: "close"},
+	})
+}
+
+// overlayHintBarOverlayTrafficCapture returns the hint bar entries for the
+// traffic-capture overlay, branching on the current capture phase. Inline
+// hints inside the overlay body are deliberately avoided — the project-wide
+// convention is the bottom-of-screen status bar.
+func (m Model) overlayHintBarOverlayTrafficCapture() string {
+	switch m.captureOverlay.phase {
+	case capturePhaseEndpointPick:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "j/k", Desc: "navigate"},
+			{Key: "Enter", Desc: "pick"},
+			{Key: "Esc", Desc: "close"},
+		})
+	case capturePhaseLive:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "s", Desc: "stop"},
+			{Key: "t", Desc: "status-only"},
+			{Key: "Y", Desc: "copy path"},
+			{Key: "/", Desc: "search"},
+			{Key: "j/k", Desc: "scroll"},
+			{Key: "ctrl+d/u", Desc: "half page"},
+			{Key: "ctrl+f/b", Desc: "page"},
+			{Key: "g/G", Desc: "oldest/live"},
+			{Key: "Esc", Desc: "stop"},
+		})
+	case capturePhaseStopped:
+		return m.renderHints([]ui.HintEntry{
+			{Key: "Enter", Desc: "restart"},
+			{Key: "e", Desc: "edit filter"},
+			{Key: "Y", Desc: "copy path"},
+			{Key: "j/k", Desc: "scroll"},
+			{Key: "ctrl+d/u", Desc: "half page"},
+			{Key: "ctrl+f/b", Desc: "page"},
+			{Key: "Esc", Desc: "close"},
+		})
+	default: // capturePhaseConfig
+		return m.renderHints([]ui.HintEntry{
+			{Key: "Enter", Desc: "start"},
+			{Key: "Tab/jk", Desc: "next field"},
+			{Key: "h/l", Desc: "cycle value"},
+			{Key: "Esc", Desc: "close"},
+		})
+	}
+}
+
+func (m Model) overlayHintBarOverlayCanI() string {
+	if m.canISearchActive {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "search"},
+			{Key: "enter", Desc: "apply"},
+			{Key: "esc", Desc: "clear"},
+		})
+	}
+	if m.canISearchQuery != "" {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "j/k", Desc: "navigate"},
+			{Key: "/", Desc: "edit search"},
+			{Key: "esc", Desc: "clear search"},
+		})
+	}
+	filterLabel := "all"
+	if m.canIAllowedOnly {
+		filterLabel = "allowed only"
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "j/k", Desc: "navigate"},
+		{Key: "a", Desc: filterLabel},
+		{Key: "s", Desc: "switch subject"},
+		{Key: "/", Desc: "search groups"},
+		{Key: "Tab", Desc: "Who-Can"},
+		{Key: "q/Esc", Desc: "close"},
+	})
+}
+
+func (m Model) overlayHintBarOverlayColumnToggle() string {
+	if m.columnToggleFilterActive {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "type", Desc: "filter"},
+			{Key: "esc", Desc: "clear/close"},
+		})
+	}
+	return m.renderHints([]ui.HintEntry{
+		{Key: "space", Desc: "toggle"},
+		{Key: "J/K", Desc: "reorder"},
+		{Key: "c", Desc: "clear"},
+		{Key: "R", Desc: "reset"},
+		{Key: "/", Desc: "filter"},
+		{Key: "Enter", Desc: "save"},
+		{Key: "Esc", Desc: "discard"},
+	})
+}
+
+// overlayHintBarOverlayRightsizing returns the hint bar entries for
+// the right-sizing overlay. Loading state shows only "esc cancel"
+// since the other actions need data; loaded state shows y/r/esc plus
+// the picker chords:
+//
+//   - `[/]: strategy` only when more than one strategy is available
+//     (otherwise the cycle is a no-op and the hint would be a lie).
+//   - `</>: headroom` always (model.RightsizingHeadrooms is a fixed
+//     6-entry list, so the cycle always has somewhere to go).
+//
+// The vim-nav set (jk/gG/ctrl+d-u/ctrl+f-b) appears only when the
+// table actually overflows the visible height.
+func (m Model) overlayHintBarOverlayRightsizing() string {
+	if m.rightsizing.loading {
+		return m.renderHints([]ui.HintEntry{
+			{Key: "esc", Desc: "cancel"},
+		})
+	}
+	hints := []ui.HintEntry{
+		{Key: "y", Desc: "copy as YAML"},
+		{Key: "r", Desc: "refresh"},
+	}
+	if len(m.rightsizing.available) > 1 {
+		hints = append(hints, ui.HintEntry{Key: "[/]", Desc: "strategy"})
+	}
+	hints = append(hints, ui.HintEntry{Key: "</>", Desc: "headroom"})
+	hints = append(hints, ui.HintEntry{Key: "esc", Desc: "close"})
+	if m.rightsizing.data != nil && len(m.rightsizing.data.Containers)*2 > rightsizingVisibleRows(m) {
+		// Match the vim-nav set used by the NetworkPolicy + Help
+		// overlays (jk single, gG top/bottom, ctrl+d/u half-page,
+		// ctrl+f/b full-page) so the right-sizing overlay feels
+		// consistent with lfk's other read-only inspection views.
+		hints = append(hints,
+			ui.HintEntry{Key: "j/k", Desc: "scroll"},
+			ui.HintEntry{Key: "g/G", Desc: "top/bottom"},
+			ui.HintEntry{Key: "ctrl+d/u", Desc: "half page"},
+			ui.HintEntry{Key: "ctrl+f/b", Desc: "page"},
+		)
+	}
+	return m.renderHints(hints)
+}
+
+// rightsizingVisibleRows is a rough estimate of how many table rows
+// fit in the overlay. Used to decide whether to surface the jk hint.
+// Mirrors the box-dim math in RenderRightsizingOverlay (75% screen
+// height minus chrome). Approximate is fine — this is a hint, not a
+// hard layout constraint.
+func rightsizingVisibleRows(m Model) int {
+	boxH := max(m.height*75/100, 12)
+	// Subtract outer padding (4), inner border (2), title (2), gap (1),
+	// grouped table header (2), header underline (1) = 12.
+	return max(boxH-12, 1)
+}
